EL-2494
Added webpack.dev.config.js which includes webpack configs for docs and
each of the 3 assets needed to run examples externally (lib, ng1, css).
Since I could not get the grunt-webpack plugin to handle multiple
configs at once, this is currently launched by `npm start`.
Added a grunt task to install the webpack-dev-server localhost
certificate on Windows.
Fixed CSS urls to be correctly relative.
Added a config.dev.json which uses the localhost assetsUrl, this will be
used when the dev server is run.
Updated the showcase to use css and ng1 from assets, so that can be
rebuilt by the dev server.
Removed the connect:plunker task.
Updated the readme.