S 60/el 2566 (#223)
* EL-2566

Replaced overflow-tooltip in the draggable card with a new directive
which only applies the tooltip. Since it will be deprecated soon I
decided not to make changes to the current directive.
Changed the card header layout to use the available width.
Fixed a lint issue in another stylesheet.

* EL-2566

Added back max-width.