EL-2589
Removed checkbox line-height CSS rule which was causing issues with the
row layout.
Added styles to the list view CSS to adjust spark label width to avoid
overflow.
Fixed a spark line alignment issue in LV2.
Changed the body-bg variable back to light grey, and updated docs style
to use white as the bg color.
Fixed the issue with the notification icon.