EL-2483
Fixed an issue caused by attempting to scroll before the target element
is in the DOM.
Removed setRendering/doneRendering/afterDoneRendering because they are
not needed with the snippets change.
Added scroll into view funcionality to the side navigation, to ensure
that the active category/section is on screen.
Cleaned up some code.