Less modifications for Darwin Micro Focus theme
Modified the Less build process to keep variables.less separate in assets. (This turned out not to be necessary for this ticket, but could be useful for other projects, e.g. parsing out colors like the customize page.)
Added variables for all library images, allowing them to be selectively replaced.