EL-2895 - Adding fixed header table directives and documentation (#503)
* Adding fixed header table directives and documentation

* Minor corrections

* Reverting change

* Review feedback

Now, if the user supplies a number for the table height, pixels will be used by default. However if a string is supplied then the user can provide their own unit of choice.

Also added gave Id column in the example a width to show how this is done and added a comment in the description about how to do it. I tried a few different ways to do try and do it automatically but they proved to be quite problematic.

* Fixing IE issue

* Reverting IE fix

* Fixing DOCTYPE

* Updatin snippets

* Updating documentation