EL-2961 - Fixing wrapper issues
Marquee wizard selector used hyphen instead of camel case.

The select table component was trying to fire a scope.digest whenever the select function is called. This function is only ever called by ng-click, which already does a digest so we were getting errors in the console about multiple digest cycles - this wasnt related to any changes to the wrappers but just discovered while testing them