Release 1.6.5
To accommodate the Stack B hosting of the site, as well as future CI builds, I have removed the explicit `assetsUrl` config setting and instead derived the URL from the active location.

This also means that plunker can load in Firefox in development mode by accessing the documentation at http://127.0.0.1:8080.