Fix for performance issue
The issue is that the interval created within `ensureScrollable` triggers a reinitialize of the jScrollPane plugin, which triggers another call to `ensureScrollable` and so on.

One fix which seems to work is to remove the call to `scrollpane.reinitialise()`. However this was explicitly added to address an issue which cannot now be identified so for safety I have left it in place. This fix instead ensures that the cascade of multiple setIntervals does not occur. Since the interval is never cancelled, the first one created should be sufficient.