EL-3871 - Fix timezone related karma test failures (#1325)
* Fix timezone related karma test failures

Tests depend on being run on a system in the +0 timezone, which will not always be the case. Stub the `getTimezoneOffset` function to always return 0 when running these tests.

* Fix timezone mismatch

The explicit timezone in these tests causes the date to be adjusted to the local timezone when parsed, and therefore (depending on the system timezone) may no longer match the expected date when displayed in the date range picker.