Fix initial timezone for date range e2e tests
The tests use the system timezone settings, but with the selenium setup this potentially involves two separate systems - the dev environment and the selenium docker container. It's quite likely that these two will have different timezone settings since the container is fixed to UTC. The change here is to use UTC/GMT explicitly in both the test app and the tests themselves.