Fix incorrect height on toolbar search placeholder
As seen in the e2e test where both toolbars are expanded, the placeholder can be the wrong height, causing the parent element to collapse. This seems to be due to the element's layout not being completely calculated at the time of measuring the effective height. A timeout appears to fix this.