Fix an issue where the notifications list fills the height of the screen
If you change dynamically from a bottom to a top `position` value, the `bottom` CSS value will not be cleared and the notification list will fill the entire height of the screen, blocking clicks.

Recalculate the bottom position, including unsetting it, when the notifications or one of the inputs change.